+++ src/toniarts/opendungeonkeeper/tools/convert/kmf/AnimControl.java
@@ -12,23 +12,23 @@ package toniarts.opendungeonkeeper.tools.convert.kmf;
  */
 public class AnimControl {
 
-    private short unknown1;
-    private short unknown2;
+    private int unknown1;
+    private int unknown2;
     private int unknown3;
 
-    public short getUnknown1() {
+    public int getUnknown1() {
         return unknown1;
     }
 
-    protected void setUnknown1(short unknown1) {
+    protected void setUnknown1(int unknown1) {
         this.unknown1 = unknown1;
     }
 
-    public short getUnknown2() {
+    public int getUnknown2() {
         return unknown2;
     }
 
-    protected void setUnknown2(short unknown2) {
+    protected void setUnknown2(int unknown2) {
         this.unknown2 = unknown2;
     }
